Revert "chore: add state type"
This reverts commit 050ba53c5fade99205149e07ba3c89e56f37a6df.

+++ x/gal/keeper/claim.go
@@ -87,7 +87,7 @@ func (k Keeper) TotalClaimableAssets(ctx sdk.Context, zone ibcstakingtypes.Regis
 	}
 
 	for _, record := range records.Records {
-		if record.State == DELEGATE_SUCCESS && record.OracleVersion < oracleVersion {
+		if record.State == int64(DELEGATE_SUCCESS) && record.OracleVersion < oracleVersion {
 			result = result.Add(*record.Amount)
 		}
 	}
